pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "peint"
version = "0.0.1"
description = "A framework for models of biological evolution"
authors = [
    {name = "Stephen Z. Lu", email = "stephen.lu@berkeley.edu"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # --------- Core ML Framework --------- #
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "torchmetrics>=0.11.4",
    "flash-attn==2.7.4.post1; sys_platform == 'linux'",
    "triton>=3.1.0; sys_platform == 'linux'",
    "fair-esm",
    "cherryml",
    # --------- Configuration Management --------- #
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    # --------- Data Science & Utilities --------- #
    "numpy==1.26.4",
    "pandas==2.1.4",
    "matplotlib==3.8.0",
    "tqdm==4.66.3",
    "biopython>=1.85",
    "scikit-learn>=1.7.1",
    "more-itertools==10.1.0",
    "typing-extensions==4.8.0",
    # --------- Jupyter & Notebooks --------- #
    "jupyter==1.0.0",
    "papermill==2.4.0",
    # --------- Experiment Tracking --------- #
    "wandb==0.15.12",
    # --------- Development Tools --------- #
    "rich",
    "rootutils",
    "torchtyping>=0.1.5",
    "black>=25.1.0",
    "peft>=0.17.1",
    "einops>=0.8.1",
    "evo",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "pytest>=7.4.2",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=3.5.0",
    "pytest>=7.4.2",
]
no-build-isolation-package = ['flash-attn', 'setuptools']

[tool.uv.sources]
fair-esm = { git = "https://github.com/facebookresearch/esm", rev = "2b369911bb5b4b0dda914521b9475cad1656b2ac" }
cherryml = { git = "https://github.com/songlab-cal/CherryML" }
evo = { workspace = true }

[tool.uv.workspace]
members = [
    "evo",
]

[tool.black]
line-length = 100
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
